Fix salary math and add rehan program
Updated the salary example to use floating-point values for basic salary, DA, HRA, and gross pay so decimal amounts are handled correctly. Added a new C program, rehan.c, that prints a message based on the number of bodies entered. Rebuilt the related binaries and debug symbols.

+++ 2.September/clg/salary.c
@@ -1,12 +1,12 @@
 #include<stdio.h>
 int main(){
-    int basic_salary,da,hra,gross_salary;
+    float basic_salary,da,hra,gross_salary;
     printf("Enter Basic Salary : ");
-    scanf("%d",&basic_salary);
+    scanf("%f",&basic_salary);
     da=basic_salary*0.4;
     hra=basic_salary*0.2;
     gross_salary=basic_salary+da+hra;
-    printf("Gross Salary : %d\n",gross_salary);
-    printf("DA : %d\n",da);
-    printf("HRA : %d\n",hra);
+    printf("Gross Salary : %f\n",gross_salary);
+    printf("DA : %f\n",da);
+    printf("HRA : %f\n",hra);
 }
